// mgpng/src/app/componentes/gestion-poa/prog-institucional/models/prog-ins.model.ts
export interface ProInsModel {
  codigo: string;
  codigoMetaFk: string;
  descripcion: string;
  estado?: string;
  alineacion?: string;
  usuarioCreacion?: string;
  fechaCreacion?: Date;
  usuarioModificacion?: string | null;
  fechaModificacion?: Date | null;
}

export interface ProInsUpdateModel {
  codigoMetaFk: string;
  descripcion: string;
  estado?: string;
  alineacion?: string;
  codigo?: string;
}

export interface TablePageEvent {
  first?: number;
  rows?: number;
}